Assignment 3 ready for submission

+++ Assignment3/src/Assignment3.java
@@ -1,7 +1,7 @@
 // CSE 110     : CSE 110 / Online
-// Assignment  : Assignment 1
+// Assignment  : Assignment 3
 // Author      : Teodoro Salgado 1220358243
-// Description :
+// Description : Program takes user inputs of share prices and determines whether to sell, buy, or hold
 import java.util.Scanner;
 
 public class Assignment3 {
@@ -17,13 +17,13 @@ public class Assignment3 {
 
 		/* PROMPT FOR AND COLLECT INPUTS */
 		System.out.print("Current Shares  : ");
-		currentShares = input.nextInt();
+		currentShares = input.nextDouble();
 		System.out.print("Purchase Price  : ");
-		purchasePrice = input.nextInt();
+		purchasePrice = input.nextDouble();
 		System.out.print("Market Price    : ");
-		marketPrice = input.nextInt();
+		marketPrice = input.nextDouble();
 		System.out.print("Available Funds : ");
-		availableFunds = input.nextInt();
+		availableFunds = input.nextDouble();
 		System.out.println();
 
 		/* COMPUTE REQUIRED VALUES */
@@ -41,9 +41,11 @@ public class Assignment3 {
 		/* DISPLAY REQUIRED OUTPUTS */
 		// Sell or hold shares if purchase price is less than market price
 		if (purchasePrice < marketPrice){
+			// Sell shares if transaction is profitable after transaction fee
 			if (totalSellValue > 10){
 				System.out.printf("Sell %d shares\n", (int)numberOfSharesToSell);
 			}
+			// Otherwise hold shares
 			else {
 				System.out.println("Hold shares");
 			}
@@ -51,9 +53,11 @@ public class Assignment3 {
 
 		// Buy or hold shares if purchase price is greater than market price
 		else if (purchasePrice > marketPrice) {
+			// Buy shares if value of the transaction is greater than the transaction fee
 			if (totalBuyValue > 10) {
 				System.out.printf("Buy %d shares\n", (int)numberOfSharesToBuy);
 			}
+			// Otherwise hold shares
 			else {
 				System.out.println("Hold shares");
 			}
